Extract killers civilian clothes to separate file
