+++ addons/equipment/CfgSerialKillers.hpp
@@ -415,322 +415,7 @@ class CfgSerialKillers
                     };
 
                     /* Clothing */
-                    class rhs_balaclava1_olive {};
-                    class G_Bandanna_tan {};
-                    class G_Bandanna_blk {};
-                    class G_Bandanna_khk {};
-                    class G_Bandanna_beast {};
-                    class G_Bandanna_aviator {};
-                    class G_Bandanna_shades {};
-                    class G_Bandanna_oli {};
-                    class G_Bandanna_sport {};
-                    class G_Lady_Blue {};
-                    class rhs_ess_black {};
-                    class G_Diving {};
-                    class G_Combat_Goggles_tna_F {};
-                    class G_Combat {};
-                    class G_I_Diving {};
-                    class G_O_Diving {};
-                    class G_B_Diving {};
-                    class G_Lowprofile {};
-                    class G_Goggles_VR {};
-                    class rhs_balaclava {};
-                    class G_Balaclava_blk {};
-                    class G_Balaclava_lowprofile {};
-                    class G_Balaclava_oli {};
-                    class G_Balaclava_combat {};
-                    class G_Balaclava_TI_blk_F {};
-                    class G_Balaclava_TI_tna_F {};
-                    class G_Balaclava_TI_G_blk_F {};
-                    class G_Balaclava_TI_G_tna_F {};
-                    class G_Squares {};
-                    class G_Squares_Tinted {};
-                    class rds_long_hair_02 {};
-                    class rds_long_hair_01 {};
-                    class rds_long_hair_03 {};
-                    class G_Respirator_white_F {};
-                    class G_Respirator_yellow_F {};
-                    class G_Respirator_blue_F {};
-                    class G_Spectacles {};
-                    class rhs_googles_orange {};
-                    class rhs_googles_black {};
-                    class rhs_googles_yellow {};
-                    class rhs_googles_clear {};
-                    class G_Aviator {};
-                    class G_EyeProtectors_F {};
-                    class G_Shades_Red {};
-                    class G_EyeProtectors_Earpiece_F {};
-                    class G_Shades_Green {};
-                    class G_Sport_Red {};
-                    class G_Sport_Blackyellow {};
-                    class G_Sport_Checkered {};
-                    class G_Sport_BlackWhite {};
-                    class G_Sport_Greenblack {};
-                    class G_Tactical_Black {};
-                    class G_Spectacles_Tinted {};
-                    class G_Tactical_Clear {};
-                    class rhsusf_shemagh_white {};
-                    class rhsusf_shemagh_tan {};
-                    class rhsusf_shemagh2_tan {};
-                    class rhsusf_shemagh2_od {};
-                    class rhsusf_shemagh2_grn {};
-                    class rhsusf_shemagh_grn {};
-                    class rhsusf_shemagh2_white {};
-                    class rhsusf_shemagh2_gogg_grn {};
-                    class rhsusf_shemagh_gogg_grn {};
-                    class rhsusf_shemagh2_gogg_od {};
-                    class rhsusf_shemagh_gogg_tan {};
-                    class rhsusf_shemagh2_gogg_tan {};
-                    class rhsusf_shemagh_gogg_white {};
-                    class rhsusf_shemagh_gogg_od {};
-                    class rhsusf_oakley_goggles_clr {};
-                    class rhsusf_oakley_goggles_blk {};
-                    class rhsusf_shemagh2_gogg_white {};
-                    class G_Shades_Black {};
-                    class G_Sport_Blackred {};
-                    class G_Shades_Blue {};
-                    class rhsusf_oakley_goggles_ylw {};
-                    class rhs_scarf {};
-                    class G_WirelessEarpiece_F {};
-                    class rds_worker_cap4 {};
-                    class rds_worker_cap3 {};
-                    class rds_worker_cap2 {};
-                    class rds_worker_cap1 {};
-                    class rds_Villager_cap1 {};
-                    class rds_Villager_cap2 {};
-                    class rds_Villager_cap3 {};
-                    class rds_Villager_cap4 {};
-                    class rds_Woodlander_cap4 {};
-                    class rds_Woodlander_cap1 {};
-                    class rds_Woodlander_cap2 {};
-                    class rds_Woodlander_cap3 {};
-                    class H_Shemag_olive_hs {};
-                    class H_Shemag_olive {};
-                    class H_ShemagOpen_tan {};
-                    class H_HeadSet_orange_F {};
-                    class H_HeadSet_red_F {};
-                    class H_HeadSet_white_F {};
-                    class H_HeadSet_black_F {};
-                    class H_HeadSet_yellow_F {};
-                    class rds_Profiteer_cap1 {};
-                    class rds_Profiteer_cap3 {};
-                    class rds_Profiteer_cap4 {};
-                    class rds_Profiteer_cap2 {};
-                    class H_HeadBandage_bloody_F {};
-                    class H_HeadBandage_clean_F {};
-                    class H_HeadBandage_stained_F {};
-                    class H_EarProtectors_red_F {};
-                    class H_EarProtectors_orange_F {};
-                    class H_EarProtectors_black_F {};
-                    class H_EarProtectors_white_F {};
-                    class H_EarProtectors_yellow_F {};
-                    class rds_rocker_hair3 {};
-                    class rds_rocker_hair2 {};
-                    class rds_rocker_hair1 {};
-                    class H_RacingHelmet_1_yellow_F {};
-                    class H_RacingHelmet_1_white_F {};
-                    class H_RacingHelmet_2_F {};
-                    class H_RacingHelmet_1_black_F {};
-                    class H_RacingHelmet_1_red_F {};
-                    class H_RacingHelmet_1_F {};
-                    class H_RacingHelmet_1_orange_F {};
-                    class H_RacingHelmet_1_blue_F {};
-                    class H_RacingHelmet_4_F {};
-                    class H_RacingHelmet_1_green_F {};
-                    class H_RacingHelmet_3_F {};
-                    class H_Helmet_Skate {};
-                    class H_Construction_headset_vrana_F {};
-                    class H_Construction_earprot_vrana_F {};
-                    class H_Construction_basic_vrana_F {};
-                    class H_Construction_headset_orange_F {};
-                    class H_Construction_earprot_orange_F {};
-                    class H_Construction_headset_red_F {};
-                    class H_Construction_basic_orange_F {};
-                    class H_Construction_headset_black_F {};
-                    class H_Construction_earprot_red_F {};
-                    class H_Construction_basic_red_F {};
-                    class H_Construction_earprot_black_F {};
-                    class H_Construction_basic_black_F {};
-                    class H_Construction_headset_white_F {};
-                    class H_Construction_earprot_white_F {};
-                    class H_Construction_earprot_yellow_F {};
-                    class H_Construction_basic_white_F {};
-                    class H_Construction_headset_yellow_F {};
-                    class H_Construction_basic_yellow_F {};
-                    class H_Booniehat_tna_F {};
-                    class H_Booniehat_mcamo {};
-                    class H_Booniehat_khk {};
-                    class H_Hat_Safari_sand_F {};
-                    class H_Hat_Safari_olive_F {};
-                    class H_StrawHat_dark {};
-                    class H_StrawHat {};
-                    class rhs_Booniehat_flora {};
-                    class rhs_Booniehat_digi {};
-                    class H_Booniehat_tan {};
-                    class H_Booniehat_oli {};
-                    class H_Hat_grey {};
-                    class H_Hat_blue {};
-                    class H_Hat_checker {};
-                    class H_Hat_tan {};
-                    class H_Hat_brown {};
-                    class H_Hat_camo {};
-                    class rhsgref_fieldcap_ttsko_urban {};
-                    class rhsgref_fieldcap_ttsko_mountain {};
-                    class rhsgref_fieldcap_ttsko_digi {};
-                    class rhsgref_fieldcap_ttsko_forest {};
-                    class H_MilCap_dgtl {};
-                    class H_MilCap_ghex_F {};
-                    class H_MilCap_tna_F {};
-                    class H_MilCap_gry {};
-                    class H_MilCap_blue {};
-                    class H_MilCap_mcamo {};
-                    class H_MilCap_ocamo {};
-                    class H_Cap_marshal {};
-                    class rhs_fieldcap_vsr {};
-                    class rhs_fieldcap_khk {};
-                    class rhs_fieldcap_ml {};
-                    class rhs_fieldcap_digi2 {};
-                    class rhs_fieldcap_digi {};
-                    class H_Watchcap_khk {};
-                    class H_Watchcap_blk {};
-                    class H_Cap_headphones {};
-                    class H_Watchcap_camo {};
-                    class H_Watchcap_cbr {};
-                    class H_Cap_brn_SPECOPS {};
-                    class H_Cap_blk_Raven {};
-                    class H_Cap_grn {};
-                    class H_Cap_tan_specops_US {};
-                    class H_Cap_khaki_specops_UK {};
-                    class rhs_fieldcap {};
-                    class H_Cap_surfer {};
-                    class H_Cap_Orange_IDAP_F {};
-                    class H_Cap_oli {};
-                    class H_Cap_oli_hs {};
-                    class H_Cap_blu {};
-                    class H_Cap_press {};
-                    class H_Cap_tan {};
-                    class H_Cap_blk_ION {};
-                    class H_Cap_red {};
-                    class H_Cap_grn_BI {};
-                    class H_Cap_blk_CMMG {};
-                    class H_Cap_usblack {};
-                    class H_Cap_White_IDAP_F {};
-                    class H_Cap_blk {};
-                    class H_Cap_Black_IDAP_F {};
-                    class rhssaf_booniehat_md2camo {};
-                    class H_Booniehat_dgtl {};
-                    class rhsgref_Booniehat_alpen {};
-                    class rhs_Booniehat_m81 {};
-                    class rhs_booniehat2_marpatd {};
-                    class rhs_Booniehat_ocp {};
-                    class rhs_Booniehat_ucp {};
-                    class rhs_booniehat2_marpatwd {};
-                    class H_Booniehat_khk_hs {};
-                    class rhssaf_booniehat_woodland {};
-                    class rhssaf_booniehat_digital {};
-                    class H_ShemagOpen_khk {};
-                    class rhssaf_bandana_digital {};
-                    class rhssaf_bandana_digital_desert {};
-                    class rhssaf_bandana_md2camo {};
-                    class rhssaf_bandana_smb {};
-                    class H_Bandanna_gry {};
-                    class rhssaf_bandana_oakleaf {};
-                    class H_Bandanna_khk {};
-                    class H_Bandanna_khk_hs {};
-                    class H_Bandanna_camo {};
-                    class H_Bandanna_mcamo {};
-                    class H_Bandanna_cbr {};
-                    class H_Bandanna_surfer {};
-                    class H_Bandanna_sand {};
-                    class H_Bandanna_sgg {};
-                    class H_Bandanna_blu {};
-                    class H_Bandanna_surfer_grn {};
-                    class rhsgref_bcap_specter {};
-                    class H_Bandanna_surfer_blk {};
-                    class rhs_beanie_green {};
-                    class rhs_beanie {};
-                    class rhssaf_beret_green {};
-                    class rhssaf_beret_para {};
-                    class rhssaf_beret_black {};
-                    class rhssaf_beret_red {};
-                    class H_Beret_gen_F {};
-                    class H_Beret_blk {};
-                    class rhssaf_beret_blue_un {};
-                    class rds_uniform_citizen1 {};
-                    class rds_uniform_assistant {};
-                    class rds_uniform_citizen2 {};
-                    class rds_uniform_citizen3 {};
-                    class rds_uniform_citizen4 {};
-                    class rds_uniform_doctor {};
-                    class rds_uniform_Functionary2 {};
-                    class rds_uniform_Functionary1 {};
-                    class U_C_ConstructionCoverall_Vrana_F {};
-                    class U_C_ConstructionCoverall_Red_F {};
-                    class U_C_ConstructionCoverall_Blue_F {};
-                    class U_C_ConstructionCoverall_Black_F {};
-                    class U_OrestesBody {};
-                    class U_I_C_Soldier_Bandit_2_F {};
-                    class U_I_C_Soldier_Bandit_5_F {};
-                    class U_I_C_Soldier_Bandit_1_F {};
-                    class U_I_C_Soldier_Bandit_4_F {};
-                    class U_I_C_Soldier_Bandit_3_F {};
-                    class U_C_Man_casual_5_F {};
-                    class U_C_Man_casual_1_F {};
-                    class U_C_man_sport_2_F {};
-                    class U_C_man_sport_1_F {};
-                    class U_BG_Guerilla1_2_F {};
-                    class U_C_man_sport_3_F {};
-                    class U_BG_Guerilla1_1 {};
-                    class U_C_Man_casual_6_F {};
-                    class U_C_Man_casual_4_F {};
-                    class U_C_Man_casual_3_F {};
-                    class U_C_Man_casual_2_F {};
-                    class rds_uniform_Profiteer1 {};
-                    class rds_uniform_priest {};
-                    class rds_uniform_Profiteer4 {};
-                    class rds_uniform_Profiteer3 {};
-                    class rds_uniform_Profiteer2 {};
-                    class rds_uniform_Rocker1 {};
-                    class rds_uniform_Rocker2 {};
-                    class rds_uniform_Rocker3 {};
-                    class rds_uniform_Rocker4 {};
-                    class rds_uniform_schoolteacher {};
-                    class U_BG_Guerilla2_2 {};
-                    class U_BG_Guerilla2_1 {};
-                    class U_BG_Guerilla2_3 {};
-                    class U_Competitor {};
-                    class U_C_IDAP_Man_shorts_F {};
-                    class U_C_IDAP_Man_casual_F {};
-                    class U_C_IDAP_Man_TeeShorts_F {};
-                    class U_C_IDAP_Man_Jeans_F {};
-                    class U_C_IDAP_Man_Tee_F {};
-                    class U_C_IDAP_Man_cargo_F {};
-                    class U_Marshal {};
-                    class U_C_Journalist {};
-                    class U_C_Mechanic_01_F {};
-                    class U_C_HunterBody_grn {};
-                    class U_BG_Guerrilla_6_1 {};
-                    class U_C_Scientist {};
-                    class U_BG_leader {};
-                    class rds_uniform_Villager1 {};
-                    class rds_uniform_Villager2 {};
-                    class rds_uniform_Villager3 {};
-                    class rds_uniform_Villager4 {};
-                    class rds_uniform_Woodlander4 {};
-                    class rds_uniform_Woodlander3 {};
-                    class rds_uniform_Woodlander2 {};
-                    class rds_uniform_Woodlander1 {};
-                    class rds_uniform_Worker4 {};
-                    class rds_uniform_Worker3 {};
-                    class rds_uniform_Worker2 {};
-                    class rds_uniform_Worker1 {};
-                    class U_C_Poor_1 {};
-                    class U_C_Poloshirt_burgundy {};
-                    class U_C_Poloshirt_salmon {};
-                    class U_C_Poloshirt_redwhite {};
-                    class U_C_Poloshirt_blue {};
-                    class U_C_Poloshirt_tricolour {};
-                    class U_C_Poloshirt_stripped {};
+                    #include "presets/cup-rhs/killers-civilian-clothing.hpp"
                 };
             };
             class Police
